microservices/qr-service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pymongo>=4.0.1
motor>=2.5.1
pydantic[email]>=2.4.2
qrcode[pil]>=7.3
Pillow>=9.0.0
requests>=2.26.0
python-magic>=0.4.24
minio>=7.1.0
httpx>=0.24.0 